вынес Interface User и Todo в отдельный компонент
