+++ src/app/users-list/user-delete-dialog/user-delete-dialog.component.html
@@ -1,8 +1,8 @@
 <div>
-    <h1>Вы действительно ходите удалить пользователя {{data.user.name}}? </h1>
+    <h1>Вы действительно ходите удалить пользователя {{userName}}? </h1>
     <div>
-        <button [mat-dialog-close]="yes">Да</button>
-        <button [mat-dialog-close]="no">Нет</button>
+        <button [mat-dialog-close]="true">Да</button>
+        <button [mat-dialog-close]="false">Нет</button>
     </div>
 </div>
